Added security on the TimeSlotController.
--HG--
extra : convert_revision : svn%3A25cf9c4d-fa41-0410-a78d-8fe958b11fb0/trunk%40802

+++ src/UI/Controllers/TimeSlotController.cs
@@ -2,6 +2,7 @@ using System;
 using System.Web.Mvc;
 using CodeCampServer.Core.Domain;
 using CodeCampServer.Core.Domain.Model;
+using CodeCampServer.Core.Services;
 using CodeCampServer.UI.Helpers.Filters;
 using CodeCampServer.UI.Helpers.Mappers;
 using CodeCampServer.UI.Models.Forms;
@@ -14,12 +15,14 @@ namespace CodeCampServer.UI.Controllers
 		private readonly ITimeSlotRepository _repository;
 		private readonly ITimeSlotMapper _mapper;
 		private readonly ISessionRepository _sessionsRepository;
+	    private readonly ISecurityContext _securityContext;
 
-		public TimeSlotController(ITimeSlotRepository repository, ITimeSlotMapper mapper, ISessionRepository sessionsRepository) : base(repository, mapper)
+	    public TimeSlotController(ITimeSlotRepository repository, ITimeSlotMapper mapper, ISessionRepository sessionsRepository, ISecurityContext securityContext) : base(repository, mapper)
 		{
 			_repository = repository;
 			_mapper = mapper;
 			_sessionsRepository = sessionsRepository;
+		    _securityContext = securityContext;
 		}
 
 		public ActionResult Edit(TimeSlot timeslot)
@@ -29,6 +32,10 @@ namespace CodeCampServer.UI.Controllers
 				TempData.Add("message", "Time slot has been deleted.");
 				return RedirectToAction<TimeSlotController>(c => c.Index(null));
 			}
+            if (!_securityContext.HasPermissionsFor(timeslot.Conference))
+            {
+                return NotAuthorizedView;
+            }
 
 			return View(_mapper.Map(timeslot));
 		}
@@ -42,6 +49,10 @@ namespace CodeCampServer.UI.Controllers
 
 		public ActionResult New(Conference conference)
 		{
+            if(!_securityContext.HasPermissionsFor(conference))
+            {
+                return NotAuthorizedView;
+            }
 			var form = new TimeSlotForm() {ConferenceId = conference.Id, ConferenceKey = conference.Key};
 			return View("edit", form);
 		}
@@ -49,6 +60,11 @@ namespace CodeCampServer.UI.Controllers
 		[ValidateModel(typeof (TimeSlotForm))]
 		public ActionResult Save([Bind(Prefix = "")] TimeSlotForm form, Conference conference,string urlreferrer)
 		{
+            if(!_securityContext.HasPermissionsFor(conference))
+            {
+                return NotAuthorizedView;
+            }
+
 			Func<TimeSlot, ActionResult> successRedirect = GetSuccessRedirect(conference, urlreferrer);
 
 			return ProcessSave(form, successRedirect);
@@ -56,6 +72,10 @@ namespace CodeCampServer.UI.Controllers
 
 		public ActionResult Delete(TimeSlot timeslot)
 		{
+            if(!_securityContext.HasPermissionsFor(timeslot.Conference))
+            {
+                return NotAuthorizedView;
+            }
 			if(_sessionsRepository.GetAllForTimeSlot(timeslot).Length==0)
 			{
 				_repository.Delete(timeslot);
